server/requirements.txt
fastapi==0.110.2
google-generativeai==0.5.2
pydantic==1.10.2
python-dotenv==1.0.0
uagents==0.11.1
requests==2.31.0